chat models fix for demo
