+++ backend/src/chat/admin.py
@@ -1,4 +1,4 @@
 from django.contrib import admin
-from .models import Message
+from patients.models import Message
 
 admin.site.register(Message)
