--- backend/src/chat/models.py
@@ -1,14 +1,0 @@
-from django.contrib.auth import get_user_model
-from django.db import models
-
-User = get_user_model()
-class Message(models.Model):
-    author = models.ForeignKey(User, on_delete=models.CASCADE)
-    content = models.TextField()
-    timestamp = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return self.author.username
-
-    def last_10_messages():
-        return Message.objects.order_by('-timestamp').all()[:10]
