+++ backend/src/patients/models.py
@@ -1,4 +1,6 @@
 from django.db import models
+from django.contrib.auth import get_user_model
+
 
 
 class Patient(models.Model):
@@ -62,3 +64,16 @@ class Report(models.Model):
 
     def __str__(self):
         return 'report for %s' % (self.appointment)
+
+
+User = get_user_model()
+class Message(models.Model):
+    author = models.ForeignKey(User, on_delete=models.CASCADE)
+    content = models.TextField()
+    timestamp = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return self.author.username
+
+    def last_10_messages():
+        return Message.objects.order_by('-timestamp').all()[:10]
